+++ CHANGELOG.rst
@@ -24,6 +24,7 @@ Improvements
   metadata to/from the file.
 - Added a flag to toggle plugin detailed_results generation to minimize 
   overhead in full processing.
+- Added support for np.ndarray as Parameter values.
 
 Bugfixes
 --------